refactor: move the hero into a separate component
